Refactor the optimization subpackage
This commit refactors the `optiland.optimization` subpackage to improve its structure and maintainability.

The main changes are:
- The optimizers are now separated into `scipy` and `torch` subdirectories inside `optiland/optimization/optimizer`.
- Each optimizer class is now in its own module.
- `glass_expert.py` has been moved into the `scipy` optimizer directory.
- The top-level `__init__.py` of the subpackage has been updated to preserve the public API.
- The tests have been updated to work with the original, unmodified `problem.py`.
